Contributed by Tomas Tollemark: * fixed some broken links in build-cross.sh * changed all long varables to uintptr_t, DWORD_PTR, int etc. * fixed PCap build issues
  The WpdPack_4_1_1.zip is missing the 64bit version of libwpcap.a. See:
  http://marc.info/?l=winpcap-users&m=128638645618841
  Tomas created this manually using the instuctions at:
  http://www.perlmonks.org/?node_id=896237

  To build needs downloading libwpcap.a from
  http://www.henrynestler.com/colinux/tools/libwpcap.a.gz

+++ src/colinux/arch/i386/cpuid.h
@@ -14,16 +14,16 @@
 
 typedef union {
 	struct {
-		unsigned long eax, ebx, ecx, edx;
+		uintptr_t eax, ebx, ecx, edx;
 	};
 	struct {
-		unsigned long highest_op;
+		uintptr_t highest_op;
 		char id_string[12];
 	};
 } cpuid_t;
 
 bool_t co_i386_has_cpuid(void);
-void co_i386_get_cpuid(unsigned long op, cpuid_t *cpuid);
-co_rc_t co_i386_get_cpuid_capabilities(unsigned long *caps);
+void co_i386_get_cpuid(uintptr_t op, cpuid_t *cpuid);
+co_rc_t co_i386_get_cpuid_capabilities(uintptr_t *caps);
 
 #endif
