+++ src/colinux/common/libc.c
@@ -8,15 +8,15 @@
  */
 
 #include <colinux/os/current/memory.h>
-
+#include <string.h>
 #include "libc.h"
 
-void *co_memset(void *s, int c, long n)
+void *co_memset(void *s, int c, int n)
 {
 	return memset(s, c, n);
 }
 
-void *co_memcpy(void *dest, const void *src, long n)
+void *co_memcpy(void *dest, const void *src, int n)
 {
 	return memcpy(dest, src, n);
 }
@@ -26,12 +26,12 @@ int co_memcmp(void *s1, void *s2, int n)
 	return memcmp(s1, s2, n);
 }
 
-void *co_memmove(void *dest, const void *src, long n)
+void *co_memmove(void *dest, const void *src, int n)
 {
 	return memmove(dest, src, n);
 }
 
-void co_bzero(void *s, long n)
+void co_bzero(void *s, int n)
 {
 	memset(s, 0, n);
 }
