+++ src/colinux/os/kernel/misc.h
@@ -13,8 +13,8 @@
 
 #include <colinux/common/common.h>
 
-extern unsigned long co_os_virt_to_phys(void *addr);
-extern co_rc_t co_os_physical_memory_pages(unsigned long *pages);
+extern uintptr_t co_os_virt_to_phys(void *addr);
+extern co_rc_t co_os_physical_memory_pages(uintptr_t *pages);
 extern co_id_t co_os_current_id(void);
 
 #endif
